requirements.txt
django-filer
django=latest
psycopg2